+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Disable sentry session tracking [#610](https://github.com/datagouv/udata-front/pull/610)
 - Use metadata_modifed_at in dataservice cache keys [#611](https://github.com/datagouv/udata-front/pull/611)
+- Update dataset params parsing to deal with tags correctly [#614](https://github.com/datagouv/udata-front/pull/614)
 
 ## 6.0.3 (2024-11-19)